% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grubbsOutliers.R
\docType{data}
\name{grubbsOutlier}
\alias{grubbsOutlier}
\title{grubbsOutlier}
\format{
A data frame with 101 rows and 1 variables:
\describe{
\item{\code{values}}{double. randomly generated values with some outliers}
}
}
\source{
Generated using the \code{rnorm} function with a mean of 50 and a standard deviation of 5. One outlier has been added into the data.
}
\usage{
grubbsOutlier
}
\description{
Randomly generated, normally distributed data that can be used with the Grubbs test for outliers.
}
\keyword{datasets}
