% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irisOutliers.R
\docType{data}
\name{irisOutliers}
\alias{irisOutliers}
\title{iris dataset with outliers}
\format{
A data frame with 150 rows and 5 variables:
\describe{
\item{\code{Sepal.Length}}{double. Sepal length of the irises.}
\item{\code{Sepal.Width}}{double. Sepal width of the irises.}
\item{\code{Petal.Length}}{double. Petal length of the irises.}
\item{\code{Petal.Width}}{double. Petal width of the irises.}
\item{\code{Species}}{integer. Species of the irises.}
}
}
\source{
iris dataset from base R with generated outliers. Outliers are generated using the \href{https://www.rdocumentation.org/packages/outForest/versions/1.0.1/topics/generateOutliers}{generateOutliers} function from the outForest package.
}
\usage{
irisOutliers
}
\description{
iris dataset from base R with generated outliers
}
\keyword{datasets}
