% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOutliers.R
\name{multiOutliers}
\alias{multiOutliers}
\title{Multivariate Outlier Detection}
\usage{
multiOutliers(data, x, y, method, minPts = 5, k = 5, threshold = 0.95, ...)
}
\arguments{
\item{data}{a data frame}

\item{method}{character, supplies the method to be used for outlier detection. Methods are LoF, kNN, mahalanobis, and iForest}

\item{minPts}{(optional) numeric, minimum points used for LoF outlier detection. Default value is 5}

\item{k}{(optional) a k value used for the kNN method of outlier detection. Default value is 5}

\item{threshold}{(optional) the threshold used for kNN outlier detection. Default value is 0.95}

\item{varlist}{a list of numeric variables}
}
\value{
indices of detected outliers, if any
}
\description{
Identifies multivariate outliers using four different methods.
}
\examples{
data(mtcars)
multiOutliers(mtcars, method="mahalanobis")
multiOutliers(mtcars, method="LoF")
multiOutliers(mtcars, method="kNN")
multiOutliers(mtcars, method="iForest")
}
