% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multiOutliers.R
\name{plot.multiOutliers}
\alias{plot.multiOutliers}
\title{Plotting Multivariate Outlier Detection}
\usage{
\method{plot}{multiOutliers}(
  data,
  varlist = names(data),
  methods = c("kNN", "LoF", "mahalanobis", "iForest"),
  minPts = 10,
  k = 5,
  threshold = 0.95,
  alpha = 0.1,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{x}{the results from a multivariate outlier detection function (e.g., kNN, iForest, mahalanobis, LoF)}
}
\value{
A formatted plot of the results of the multivariate outlier detection
}
\description{
Plots identified multivariate outliers using either kNN, iForest, mahalanobis, LoF
}
\examples{
data(iris)
results <- multiOutliers(iris, method="iForest")
plot(results)
}
