% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.uniOutliers.R
\name{plot.uniOutliers}
\alias{plot.uniOutliers}
\title{Plotting Univariate Outlier Detection}
\usage{
\method{plot}{uniOutliers}(x, ...)
}
\arguments{
\item{x}{A list containing the results from a univariate outlier detection function (e.g., \code{univOutliers}).}

\item{...}{not used}
}
\value{
A plot object specific to the outlier detection method used (boxplot, MAD, or Grubbs' test).
}
\description{
Creates visualizations for univariate outlier detection using method-specific plot styles.
}
\examples{
object <- uniOutliers(data = mtcarsOutliers, x = "wt", method = "mad")
plot(object)

object2 <- uniOutliers(data = mtcarsOutliers, x = "wt", method = "boxplot")
plot(object2)

object3 <- uniOutliers(data = mtcarsOutliers, x = "wt", method = "grubbs")
plot(object3)
}
