% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.multiOutliers.R, R/print.univOutliers.R
\name{print.multiOutliers}
\alias{print.multiOutliers}
\title{Printing Multivariate outlier detection}
\usage{
\method{print}{multiOutliers}(x, ...)

\method{print}{multiOutliers}(x, ...)
}
\arguments{
\item{x}{the results from a univariate outlier detection function (e.g., \code{boxplot}, \code{grubbsTest}, or MAD)}
}
\value{
A formatted print of the results of the "multiOutliers" function

A formatted print of the results of the univariate outlier detection
}
\description{
Prints identified multivariate outliers

Prints identified univariate outliers using boxplot, Grubbs test, or MAD method
}
\examples{
multiOutliers(mtcars, method="mahalanobis")

data(mtcars)
outliers_boxplot <- univOutliers(mtcars, "mpg", method="boxplot")
print.univariateOutliers(outliers_boxplot)
outliers_mad <- univOutliers(mtcars$mpg, method="mad")
print.univariateOutliers(outliers_mad)
outliers_grubbs <- grubbs.test(mtcars$mpg)
print.univariateOutliers(outliers_grubbs)

}
