% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.univOutliers.R
\name{print.univOutliers}
\alias{print.univOutliers}
\title{Printing Univariate Outlier Detection}
\usage{
\method{print}{univOutliers}(results, method)
}
\arguments{
\item{x}{The results from a univariate outlier detection function (e.g., \code{univOutliers})}
}
\value{
A formatted print of the results of the univariate outlier detection
}
\description{
Prints identified univariate outliers using boxplot, Grubbs test, or MAD method
}
\examples{
set.seed(42)
data1 <- data.frame(
  var1 = rnorm(100, mean = 50, sd = 10),
  var2 = c(rnorm(97, mean = 50, sd = 10), 150, 160, 170),
  var3 = sample(1:100, 100, replace = TRUE, prob = (1:100)^2))

results <- univOutliers(data1, method = "boxplot")
results <- univOutliers(data1, "var1", method = "boxplot")
results <- univOutliers(data1, method = "grubbs")
results <- univOutliers(data1, "var1", method = "grubbs")
results <- univOutliers(data1, method = "mad")
results <- univOutliers(data1, "var1", method = "mad")
}
