% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qacOutliers-package.R
\docType{package}
\name{qacOutliers-package}
\alias{qacOutliers}
\alias{qacOutliers-package}
\title{qacOutliers: Identification of Multivariate and Univariate Outliers}
\description{
The goal of qacOutliers is to detect multivariate and univariate outliers using seven different methods. This package consolidates methods from multiple different R packages and makes them more easily accessible.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chenning2011/qacOutliers}
  \item Report bugs at \url{https://github.com/chenning2011/qacOutliers/issues}
}

}
\author{
\strong{Maintainer}: Caleb Henning \email{chenning@wesleyan.edu}

Authors:
\itemize{
  \item Larissa Xu \email{lxu01@wesleyan.edu}
  \item Angelica Crown \email{acrown@wesleyan.edu}
  \item Ernie Little \email{elittle@wesleyan.edu}
  \item Braeden Falzarano \email{bfalzarano@wesleyan.edu}
  \item Ral Reyes \email{rreyes@wesleyan.edu}
  \item Tegh Singh \email{tsingh@wesleyan.edu}
}

}
\keyword{internal}
