% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniOutliers.R
\name{uniOutliers}
\alias{uniOutliers}
\title{Univariate Outlier Detection}
\usage{
uniOutliers(data, x = NULL, method = "boxplot")
}
\arguments{
\item{data}{a data frame containing the variable(s) to be analyzed}

\item{x}{character, specifies the column(s) to assess for outliers. If NULL, all numeric columns are used}

\item{method}{character, the method used for outlier detection: "boxplot", "mad", or "grubbs" (default is "boxplot")}
}
\value{
A list with detected outliers, their indices, the method used, and the original dataset
}
\description{
Detects univariate outliers using boxplot, MAD, or Grubbs' test.
}
\examples{

#Example usage:
object <- uniOutliers(data = mtcarsOutliers, x="wt", method = "mad")
print(object)  # Printing the outliers and methods

object2 <- uniOutliers(data = mtcarsOutliers, x= "wt", method = "boxplot")
print(object2)  # Printing the outliers and methods

object3 <- uniOutliers(data = grubbsOutlier, method = "grubbs")
print(object3)  # Printing the outliers and methods
}
