% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univOutliers 2.R, R/univOutliers(1).R,
%   R/univOutliers.R
\name{univOutliers}
\alias{univOutliers}
\title{Univariate Outlier Detection}
\usage{
univOutliers(data, x = NULL, method = "boxplot")

univOutliers(data, x = NULL, method = "boxplot")

univOutliers(data, x = NULL, method = "boxplot")
}
\arguments{
\item{data}{A data frame containing the variable to be analyzed.}

\item{x}{A character string naming the numeric variable to assess for outliers.}

\item{method}{character, method for univariate outlier detection. Methods include boxplot, mad. See pacakage site for more information on these methods.}
}
\value{
A list containing the ggplot object and boxplot statistics.

A list containing the ggplot object and boxplot statistics.

A list containing the ggplot object and boxplot statistics.
}
\description{
Generate a boxplot specifically for univariate outlier detection and return associated statistics from base R.

Generate a boxplot specifically for univariate outlier detection and return associated statistics from base R.

Provides three methods for detecting univariate outliers in a dataset.
}
\examples{
univOutliers(mtcars, "mpg", method="boxplot")
univOutliers(mtcars, "hp", method="boxplot")
data <- c(10, 12, 10, 11, 13, 100, 10, 9, 11) # Example data
univOutliers(data, method="mad")
data2 <- Attacks$age
univOutliers(data2, method="mad")
univOutliers(mtcars, "mpg", method="boxplot")
univOutliers(mtcars, "hp", method="boxplot")
data <- c(10, 12, 10, 11, 13, 100, 10, 9, 11) # Example data
univOutliers(data, method="mad")
data2 <- Attacks$age
univOutliers(data2, method="mad")
univOutliers(mtcars, "mpg", method="boxplot")
univOutliers(mtcars, "hp", method="boxplot")
data <- c(10, 12, 10, 11, 13, 100, 10, 9, 11) # Example data
univOutliers(data, method="mad")
data2 <- Attacks$age
univOutliers(data2, method="mad")
}
