% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irisOutliers.R
\docType{data}
\name{irisOutliers}
\alias{irisOutliers}
\title{iris dataset with outliers}
\format{
A data frame with 150 rows and 5 variables:
\describe{
\item{\code{Sepal.Length}}{double. DESCRIPTION.}
\item{\code{Sepal.Width}}{double. DESCRIPTION.}
\item{\code{Petal.Length}}{double. DESCRIPTION.}
\item{\code{Petal.Width}}{double. DESCRIPTION.}
\item{\code{Species}}{integer. DESCRIPTION.}
}
}
\source{
iris dataset from base R with generated outliers
}
\usage{
irisOutliers
}
\description{
iris dataset from base R with generated outliers
}
\keyword{datasets}
