% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiOutliers.R
\name{multiOutliers}
\alias{multiOutliers}
\title{Multivariate Outlier Detection}
\usage{
multiOutliers(
  data,
  varlist = names(data),
  method = "lof",
  minPts = 10,
  k = 5,
  threshold = 0.95,
  alpha = 0.1,
  ntrees = 100,
  n = 5,
  na.rm = TRUE
)
}
\arguments{
\item{data}{a data frame}

\item{varlist}{a list of variables. The variables can be factor or numeric. Categorical variables are dropped when using the mahalanobis and kNN methods.}

\item{method}{character, supplies the method to be used for outlier detection. Methods are LoF, kNN, mahalanobis, and iForest.}

\item{minPts}{(optional) numeric, minimum points used for LoF outlier detection. Default value is 5}

\item{k}{(optional) a k value used for the kNN method of outlier detection. Default value is 5}

\item{threshold}{(optional) the threshold used for kNN outlier detection. Default value is 0.95}

\item{alpha}{(optional) the alpha used for mahalanobis distance outlier detection. Default value is 0.1}

\item{ntrees}{(optional) the number of trees used in iForest outlier detection. Default value is 100}

\item{n}{(optional) the number of points to take as outliers in iForest outlier detection. Default value is 5}

\item{na.rm}{(optional) logical, specifies whether to remove NA values. Defaults to TRUE. If the value is TRUE, removes missing data through listwise deletion.}
}
\value{
method used, dataset used, variables used for outliers detected, indices of any detected outliers, scores for the outliers, and values for optional parameters
}
\description{
Identifies multivariate outliers using four different methods.
}
\examples{
multiOutliers(mtcarsOutliers, method="mahalanobis", alpha=0.1)
multiOutliers(mtcarsOutliers, method="lof", minPts=5)
multiOutliers(mtcarsOutliers, method="knn", k=5, threshold=.95)
multiOutliers(mtcarsOutliers, method="iforest", ntrees = 50)
}
