% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis.R, R/multiOutliers(1).R,
%   R/multiOutliers.R
\name{multiOutliers}
\alias{multiOutliers}
\title{Multivariate outlier detection}
\usage{
multiOutliers(data, x, y, method, minPts, k = 5, threshold = 0.95, ...)

multiOutliers(data, x, y, method, minPts, k = 5, threshold = 0.95, ...)

multiOutliers(data, x, y, method, minPts, k = 5, threshold = 0.95, ...)
}
\arguments{
\item{data}{a data frame}

\item{x}{a numeric variable}

\item{y}{a numeric variable}

\item{method}{character, supplies the method to be used for outlier detection}

\item{minPts}{numeric, minimum points used for LoF outlier detection}

\item{k}{a k value used for the kNN method of outlier detection}

\item{threshold}{the threshold used for kNN outlier detection}

\item{varlist}{a list of numeric variables}
}
\value{
a tibble with n, mean, and standard deviation

indices of detected outliers, if any

indices of detected outliers, if any
}
\description{
Create scatter plot with linear line of best fit, correlation, and p-value

Identifies multivariate outliers

Identifies multivariate outliers using four different methods.
}
\examples{
multiOutliers(mtcars)

multiOutliers(mtcars, hisp, cyl, method="mahalanobis")

multiOutliers(mtcars, method="mahalanobis")
multiOutliers(mtcars, method="LoF")
mutliOutliers(mtcars, method="kNN")
multiOutliers(mtcars, mthod="iForest")
}
