% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahalanobis.R, R/multiOutliers 2.R,
%   R/multiOutliers(2).R, R/multiOutliers.R
\name{multiOutliers}
\alias{multiOutliers}
\title{Multivariate outlier detection}
\usage{
multiOutliers(
  data,
  varlist,
  method,
  minPts = 5,
  k = 5,
  threshold = 0.95,
  na.rm = TRUE,
  ...
)

multiOutliers(
  data,
  varlist,
  method,
  minPts = 5,
  k = 5,
  threshold = 0.95,
  na.rm = TRUE,
  ...
)

multiOutliers(
  data,
  varlist,
  method,
  minPts = 5,
  k = 5,
  threshold = 0.95,
  na.rm = TRUE,
  ...
)

multiOutliers(
  data,
  varlist,
  method,
  minPts = 5,
  k = 5,
  threshold = 0.95,
  na.rm = TRUE,
  ...
)
}
\arguments{
\item{data}{a data frame}

\item{varlist}{a list of numeric variables}

\item{method}{character, supplies the method to be used for outlier detection. Methods are LoF, kNN, mahalanobis, and iForest}

\item{minPts}{(optional) numeric, minimum points used for LoF outlier detection. Default value is 5}

\item{k}{(optional) a k value used for the kNN method of outlier detection. Default value is 5}

\item{threshold}{(optional) the threshold used for kNN outlier detection. Default value is 0.95}

\item{x}{a numeric variable}

\item{y}{a numeric variable}
}
\value{
a tibble with n, mean, and standard deviation

indices of detected outliers, if any

indices of detected outliers, if any

indices of detected outliers, if any
}
\description{
Create scatter plot with linear line of best fit, correlation, and p-value

Identifies multivariate outliers using four different methods.

Identifies multivariate outliers

Identifies multivariate outliers using four different methods.
}
\examples{
multiOutliers(mtcars)

data(mtcars)
multiOutliers(mtcars, method="mahalanobis")
multiOutliers(mtcars, method="LoF")
multiOutliers(mtcars, method="kNN")
multiOutliers(mtcars, method="iForest")
multiOutliers(mtcars, hisp, cyl, method="mahalanobis")

data(mtcars)
multiOutliers(mtcars, method="mahalanobis")
multiOutliers(mtcars, method="LoF")
multiOutliers(mtcars, method="kNN")
multiOutliers(mtcars, method="iForest")
}
