% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.multiOutliers.R
\name{plot.multiOutliers}
\alias{plot.multiOutliers}
\title{Plotting Multivariate Outlier Detection}
\usage{
\method{plot}{multiOutliers}(x, ...)
}
\arguments{
\item{x}{the results from a multivariate outlier detection function (e.g., kNN, iForest, mahalanobis, LoF)}

\item{...}{not used}
}
\value{
A formatted plot of the results of the multivariate outlier detection. Row indices are on the x-axis and ou
}
\description{
Plots identified multivariate outliers using either kNN, iForest, mahalanobis, LoF
}
\examples{
results <- multiOutliers(irisOutliers, method="iforest")
plot(results)

results <- multiOutliers(irisOutliers, method = "knn")
plot(results)

results <- multiOutliers(irisOutliers, method = "lof")
plot(results)

results <- multiOutliers(irisOutliers, method = "mahalanobis")
plot(results)

plot(multiOutliers(irisOutliers, method = "mahalanobis"))
}
