% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.multiOutliers.R
\name{print.multiOutliers}
\alias{print.multiOutliers}
\title{Printing Multivariate outlier detection}
\usage{
\method{print}{multiOutliers}(x, ...)
}
\arguments{
\item{x}{the results from "multiOutliers" function}
}
\value{
A formatted print of the results of the "multiOutliers" function
}
\description{
Prints identified multivariate outliers
}
\examples{
multiOutliers(mtcars, method="mahalanobis")

}
