% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.univOutliers.R
\name{print.univOutliers}
\alias{print.univOutliers}
\title{Printing Univariate Outlier Detection}
\usage{
\method{print}{univOutliers}(x, ...)
}
\arguments{
\item{x}{the results from a univariate outlier detection function (e.g., \code{boxplot}, \code{grubbsTest}, or MAD)}
}
\value{
A formatted print of the results of the univariate outlier detection
}
\description{
Prints identified univariate outliers using boxplot, Grubbs test, or MAD method
}
\examples{
# Example usage:
object <- univOutliers(data = mtcars, method = "mad")
plot.univOutliers(object, mtcars)  # Plotting the object with 'mtcars' dataset
print(object)  # Printing the outliers and methods

object <- univOutliers(data = mtcars, method = "boxplot")
plot.univOutliers(object, mtcars)  # Plotting the object with 'mtcars' dataset
print(object)  # Printing the outliers and methods

object <- univOutliers(data = mtcars, method = "grubbs")
plot.univOutliers(object, mtcars)  # Plotting the object with 'mtcars' dataset
print(object)  # Printing the outliers and methods
}
