% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.univOutliers.R
\name{print.univOutliers}
\alias{print.univOutliers}
\title{Printing Univariate Outlier Detection}
\usage{
\method{print}{univOutliers}(x, ...)
}
\arguments{
\item{x}{the results from a univariate outlier detection function (e.g., \code{boxplot}, \code{grubbsTest}, or MAD)}
}
\value{
A formatted print of the results of the univariate outlier detection
}
\description{
Prints identified univariate outliers using boxplot, Grubbs test, or MAD method
}
\examples{
data(mtcars)
outliers_boxplot <- univOutliers(mtcars, "mpg", method="boxplot")
outliers_boxplot
outliers_mad <- univOutliers(mtcars$mpg, method="mad")
outliers_mad
outliers_grubbs <- univOutliers(mtcars, "mpg", method="grubbs")
outliers_grubbs

}
