% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univOutliers.R
\name{univOutliers}
\alias{univOutliers}
\title{Univariate Outlier Detection}
\usage{
univOutliers(data, x = NULL, method = "boxplot")
}
\arguments{
\item{data}{A data frame containing the variable to be analyzed.}

\item{x}{A character string naming the numeric variable to assess for outliers.}

\item{method}{A character supplying the method used for outlier detection. Methods are boxplot, mad, and grubbs.}
}
\value{
A list containing the ggplot object and boxplot statistics.
}
\description{
Provides three methods for detecting univariate outliers in a dataset.
}
\examples{
# Boxplot Method on test data
univOutliers(data1, "normal_data", method = "boxplot")
univOutliers(data1, "with_outlier", method = "boxplot")
univOutliers(data1, "multiple_outliers", method = "boxplot")

# MAD Method on test data
univOutliers(data1, "normal_data", method = "mad")
univOutliers(data1, "with_outlier", method = "mad")
univOutliers(data1, "multiple_outliers", method = "mad")

# Grubbs' Method on test data
univOutliers(data1, "normal_data", method = "grubbs")
univOutliers(data1, "with_outlier", method = "grubbs")
univOutliers(data1, "multiple_outliers", method = "grubbs")
}
