% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univOutliers.R
\name{univOutliers}
\alias{univOutliers}
\title{Univariate Outlier Detection}
\usage{
univOutliers(data, x = NULL, method = "boxplot")
}
\arguments{
\item{data}{a data frame containing the variable(s) to be analyzed}

\item{x}{character, specifies the column(s) to assess for outliers. If NULL, all numeric columns are used}

\item{method}{character, the method used for outlier detection: "boxplot", "mad", or "grubbs" (default is "boxplot")}
}
\value{
A list with detected outliers, their indices, the method used, and the original dataset
}
\description{
Detects univariate outliers using boxplot, MAD, or Grubbs' test.
}
\examples{
Example usage:
object <- univOutliers(data = mtcars, x="carb", method = "mad")
plot.univOutliers(object)  # Plotting the object with 'mtcars' dataset
print(object)  # Printing the outliers and methods

object2 <- univOutliers(data = mtcars, x= "wt", method = "boxplot")
plot.univOutliers(object2)  # Plotting the object with 'mtcars' dataset
print(object2)  # Printing the outliers and methods

object3 <- univOutliers(data = mtcars, x = "carb", method = "grubbs")
plot.univOutliers(object3)  # Plotting the object with 'mtcars' dataset
print(object3)  # Printing the outliers and methods
}
