% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univOutliers.R
\name{univOutliers}
\alias{univOutliers}
\title{Univariate Outlier Detection}
\usage{
univOutliers(data, x = NULL, method = "boxplot")
}
\arguments{
\item{data}{A data frame containing the variable to be analyzed.}

\item{x}{A character string naming the numeric variable to assess for outliers.}

\item{method}{A character supplying the method used for outlier detection. Methods are boxplot, mad, and grubbs.}
}
\value{
A list containing the ggplot object and boxplot statistics.
}
\description{
Provides three methods for detecting univariate outliers in a dataset.
}
\examples{
# Generate example data
set.seed(42)
library(EnvStats)
data1 <- data.frame(
  var1 = rnorm(100, mean = 50, sd = 10),
  var2 = c(rnorm(97, mean = 50, sd = 10), 150, 160, 170),
  var3 = sample(1:100, 100, replace = TRUE, prob = (1:100)^2))


# Boxplot Method on test data
univOutliers(data1, "var1", method = "boxplot")
univOutliers(data1, "var2", method = "boxplot")
univOutliers(data1, method = "boxplot")

# MAD Method on test data
univOutliers(data1, "var1", method = "mad")
univOutliers(data1, "var2", method = "mad")
univOutliers(data1, method = "mad")

# Grubbs' Method on test data
univOutliers(data1, "var1", method = "grubbs")
univOutliers(data1, "var2", method = "grubbs")
univOutliers(data1, "var3", method = "grubbs")
}
