% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univOutliers.R
\name{univOutliers}
\alias{univOutliers}
\title{Univariate Outlier Detection}
\usage{
univOutliers(data, x = NULL, method = "boxplot")
}
\arguments{
\item{data}{A data frame containing the variable to be analyzed.}

\item{x}{A character string naming the numeric variable to assess for outliers.}

\item{method}{A character supplying the method used for outlier detection. Methods are boxplot, mad, and grubbs.}
}
\value{
A list containing the ggplot object and boxplot statistics.
}
\description{
Provides three methods for detecting univariate outliers in a dataset.
}
\examples{
data(mtcars)
univOutliers(mtcars, "mpg", method="boxplot")
univOutliers(mtcars, "hp", method="boxplot")
data <- c(10, 12, 10, 11, 13, 100, 10, 9, 11) # Example data
univOutliers(data, method="mad")
data2 <- Attacks$age
univOutliers(data2, method="mad")
}
